Move HTTP date processing to utils module.
